✨ Feat: 사용자 데이터 비동기 로딩 및 검색 기능 개선, 탭에 표시할 개수 계산 로직 통합
