+++ src/pages/group/PresetSettingPage.tsx
@@ -2,57 +2,28 @@ import Header from '@/components/common/header/Header';
 import PresetTab from '@/components/common/tab/PresetTab/PresetTab';
 import Tab from '@/components/common/tab/Tab/Tab';
 import PresetAddInput from '@/components/setting/presetSetting/PresetAddInput';
-import { PresetDefault, PresetTabName } from '@/constants';
-import { convertTabNameToChargers } from '@/utils/convertUtils';
+import { PresetTabName } from '@/constants';
 import usePresetSetting from '@/hooks/usePresetSetting';
 import usePresetSettingStore from '@/store/usePresetSettingStore';
 import MetaTags from '@/components/common/metaTags/MetaTags';
 import { useParams } from 'react-router-dom';
-import { useEffect, useMemo, useState } from 'react';
 import SearchInput from '@/components/common/search/SearchInput';
-import { getAllCategoryList } from '@/services/preset';
 
 const PresetSettingPage = () => {
   const { categoryList, activeTab, cateActiveTab, deleteButtonStates, presetData } =
     usePresetSettingStore();
-  const { handleSelectClick, handleDeleteClick, handleBack, handleTabChange, handleCateTabChange } =
-    usePresetSetting();
+  const {
+    handleSelectClick,
+    handleDeleteClick,
+    handleBack,
+    handleTabChange,
+    handleCateTabChange,
+    searchQuery,
+    setSearchQuery,
+    chargers,
+    userData,
+  } = usePresetSetting();
   const { channelId } = useParams();
-  const [searchQuery, setSearchQuery] = useState('');
-  const [userData, setUserData] = useState(presetData);
-  const [presetDefaultData] = useState(PresetDefault);
-
-  useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const response = await getAllCategoryList({ channelId: Number(channelId) });
-        setUserData(response.result.presetCategoryList);
-      } catch (error) {
-        console.error('프리셋 리스트 조회 오류: ', error);
-      }
-    };
-    fetchUserData();
-  }, [channelId]);
-
-  // 검색 결과 개수 계산
-  const getFilteredCount = (data: typeof presetData) => {
-    return data.flatMap(category =>
-      category.presetItemList.filter(item =>
-        item.name.toLowerCase().includes(searchQuery.toLowerCase())
-      )
-    ).length;
-  };
-
-  // 탭에 표시할 개수
-  const chargers = useMemo(() => {
-    return convertTabNameToChargers(PresetTabName).map(tab => ({
-      ...tab,
-      count:
-        tab.name === PresetTabName.PRESET_DATA
-          ? getFilteredCount(presetDefaultData)
-          : getFilteredCount(userData),
-    }));
-  }, [presetDefaultData, userData, searchQuery]);
 
   return (
     <div className={`flex h-screen flex-col`}>
